refactor(modo-upload): split FileStorage, StoredFile, and utils into separate files
Moves `FileStorage` trait and `StoredFile` struct from `storage/mod.rs`
into a new `storage/types.rs`, utility functions (`ensure_within`,
`generate_filename`, `validate_logical_path`) into `storage/utils.rs`,
and the `storage()` factory into `storage/factory.rs`, keeping
`mod.rs` as imports-only per convention. Bumps to v0.1.2.
