+++ modo-upload/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "modo-upload"
-version = "0.1.1"
+version = "0.1.2"
 edition = "2024"
 license.workspace = true
 repository.workspace = true
